refactored the base code to implement a chain of responsability design pattern for better code architecture
