+++ src/parsers/parser.definition.ts
@@ -1,0 +1,6 @@
+import { Log, TransactionResponse } from "ethers";
+
+export interface LogParser {
+    platform:string
+    parseLogs(transaction:TransactionResponse, logs:ReadonlyArray<Log>, tokenId:string):number|undefined
+}
